+++ dwolla-sdk/src/main/scala/dwolla/sdk/SprayClientDwollaApi.scala
@@ -8,8 +8,9 @@ import spray.http._
 import spray.can.Http.HostConnectorSetup
 import spray.http.HttpResponse
 import spray.http.HttpHeaders.{`User-Agent`, Accept}
-import dwolla.sdk.Responses._
 import dwolla.sdk.Requests._
+import dwolla.sdk.Requests.AccountType._
+import dwolla.sdk.Responses._
 import spray.json._
 import dwolla.sdk.DwollaApiAuthResponseJsonProtocol._
 import dwolla.sdk.DwollaApiRequestJsonProtocol._
@@ -88,9 +89,16 @@ private[sdk] class SprayClientDwollaApi(settings: Option[DwollaApiSettings] = No
     executeTo(Get(uriWithQuery), mapResponse[GetBalanceResponse])
   }
 
+  def addFundingSource(accessToken: String, accountNumber: String, routingNumber: String, accountType: AccountType,
+                       name: String): Future[AddFundingSourceResponse] = {
+    val uri = Uri("/oauth/rest/fundingsources/")
+    val raw = AddFundingSourceRequest2(accessToken, accountNumber, routingNumber, accountType, name)
+    executeTo(Post(uri, raw), mapResponse[AddFundingSourceResponse])
+  }
+
   def addFundingSource(accessToken: String, accountNumber: String, routingNumber: String, accountType: String,
                        name: String): Future[AddFundingSourceResponse] = {
-    val uri = Uri("/oauth/rest/fundingsources")
+    val uri = Uri("/oauth/rest/fundingsources/")
     val raw = AddFundingSourceRequest(accessToken, accountNumber, routingNumber, accountType, name)
     executeTo(Post(uri, raw), mapResponse[AddFundingSourceResponse])
   }